Added uw_file_io app and moved file viewing/downloading into it
