+++ uw_forms/templates/uw_forms/file_row.html
@@ -2,7 +2,7 @@
 
 <tr class="file-row {% if form.errors %}danger{% endif %} {% if not form.instance.to_display %}disabled{% endif %}">
   <td class="file-description">
-    <a href="{% url 'uw_inventory_system.views.file_view' form.instance.file_field.name %}">
+    <a href="{% url 'uw_file_io.views.file_view' form.instance.file_field.name %}">
       {{ form.instance.get_name_display|filename }}
     </a>
   </td>
@@ -17,7 +17,7 @@
   {% endif %}
   <td class="file-menu">
     <i class="file-download"
-       data-target="{% url 'uw_inventory_system.views.file_download' form.instance.id %}">
+       data-target="{% url 'uw_file_io.views.file_download' form.instance.id %}">
     </i>
     {% if can_edit %}
       <i class="file-edit"></i>
