+++ uw_inventory/templates/uw_inventory/detail.html
@@ -54,7 +54,7 @@
       <div class="col-xs-8 field-container">
         <p class="form-control-static" id="inputSopFile">
           {% if formsets.sop.forms %}
-            <a href="{% url 'uw_inventory_system.views.file_view' formsets.sop.forms.0.instance.file_field.name %}">
+            <a href="{% url 'uw_file_io.views.file_view' formsets.sop.forms.0.instance.file_field.name %}">
               {{ formsets.sop.forms.0.instance.get_name_display|filename }}
             </a>
           {% endif %}
