+++ uw_inventory/templates/uw_inventory/list.html
@@ -86,7 +86,7 @@
             <tr {% if not item.to_display %}class="disabled"{% endif %}>
               <td class="sop-cell">
                 {% if item.sop_file %}
-                  <a href="{% url 'uw_inventory_system.views.file_view' item.sop_file.file_field.name %}">
+                  <a href="{% url 'uw_file_io.views.file_view' item.sop_file.file_field.name %}">
                     {% contexual_file_icon item.sop_file %}
                   </a>
                 {% endif %}
