+++ uw_inventory_system/urls.py
@@ -11,12 +11,9 @@ urlpatterns = patterns(
     url(r'^$', lambda r: HttpResponseRedirect('list/')),
     url(r'^admin/', include(admin.site.urls)),
     url(r'^list/', include('uw_inventory.urls')),
-    # url(r'^accounts/login/$', 'django_cas.views.login'),
-    # url(r'^accounts/logout/$', 'django_cas.views.logout'),
     url(r'^accounts/', include('uw_users.urls')),
     # TODO: remove this URL for production
     url(r'^forbidden/$', views.permission_denied),
-    url(r'^download/(?P<file_id>\d+)/$', views.file_download),
-    url(r'^(?P<file_name>files/.+)/$', views.file_view),
+    url(r'^files/', include('uw_file_io.urls')),
 )
 urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
