+++ uw_inventory_system/views.py
@@ -1,36 +1,5 @@
-import os
-
-from django.http import HttpResponse
 from django.shortcuts import render
-
-from uw_inventory.models import ItemFile
 
 
 def permission_denied(request):
     return render(request, 'errors/403.html')
-
-
-def file_download(request, file_id):
-    if request.method == 'GET':
-        file = ItemFile.objects.get(id=file_id).file_field
-        response = HttpResponse(file)
-        response['Content-Disposition'] = 'attachment; filename="{0}"'.format(
-            os.path.basename(file.name)
-        )
-
-        return response
-
-
-def file_view(request, file_name):
-    if request.method == 'GET':
-        file_obj = ItemFile.objects.get(file_field=file_name)
-        with open(file_name, 'r') as file:
-            response = HttpResponse(
-                file.read(),
-                content_type=file_obj.mimetype
-            )
-            response['Content-Disposition'] = 'inline; filename="{0}"'.format(
-                os.path.basename(file.name)
-            )
-            return response
-        file.closed
